pyproject.toml
[tool.poetry]
name = "fastapi-gcp-tasks"
version = "0.1.0"
description = "Trigger delayed Cloud Tasks from FastAPI"
readme = "README.md"
license = "MIT"
authors = ["Team Simplify <oss@simplify.jobs>"]

[tool.poetry.dependencies]
python = ">=3.11, <3.14"
google-cloud-tasks = ">=2.16.3, <2.20.0"
google-cloud-scheduler = ">=2.13.3, <2.20.0"
fastapi = ">=0.110.0, <0.120.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.6"
uvicorn = "^0.29.0"
mypy = "^1.10.0"
types-ujson = "^5.10.0.20240515"
types-protobuf = "^5.26.0.20240422"

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 120
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "N", # pep8-naming
    "I", # isort
    "C", # flake8-comprehensions
    "D", # pydocstyle
    "B", # flake8-bugbear
    "BLE", # flake8-blind-except
    "T100", # flake8-debugger
    "PT", # flake8-pytest
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
    "D104", # missing docstring in public package
    "D107", # missing docstring in __init__
    "D202", # no blank lines allowed after function docstring
    "D203", # 1 blank line required before class docstring
    "D212", # multi-line docstring summary should start at the first line
    "D401", # first line of docstring should be in imperative mood
    "D404", # first word of the docstring should not be "This"
    "D100", # missing docstring in public module
    "N805", # first argument of a method should be named self
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"examples/*" = ["D103"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"


[build-system]
requires = ["poetry-core>=1.8.2"]
build-backend = "poetry.core.masonry.api"
